pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "wechat-agent"
version = "0.1.0"
description = "CLI aggregator and recommender for WeChat subscription articles"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12,<1.0",
  "rich>=13.7",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "httpx>=0.27",
  "feedparser>=6.0",
  "openai>=1.66",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-mock>=3.14",
  "ruff>=0.9",
]

[project.scripts]
wechat-agent = "wechat_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/wechat_agent"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
